fix: add password protection in frontend side for delete group data

+++ src/stores/group.js
@@ -46,9 +46,15 @@ export const useGroupStore = defineStore('group', {
       }
     },
 
-    async deleteGroup(id) {
+    async deleteGroup(id, password) {
       try {
+        if (password !== "adminrdzth123") {
+          message.error("Wrong password")
+          return
+        }
+
         const response = await api.delete(`/group/${id}`)
+
         if (response.status === 200) {
           message.success('Delete group data success')
         }
